Adding texture

+++ src/color.c
@@ -6,7 +6,7 @@
 /*   By: maxpelle <maxpelle@student.42quebec.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/31 07:39:51 by eguefif           #+#    #+#             */
-/*   Updated: 2023/12/14 10:17:17 by eguefif          ###   ########.fr       */
+/*   Updated: 2023/12/14 11:53:31 by eguefif          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ t_vector	get_rgb_vect(unsigned int color)
 {
 	t_vector	rgba;
 
-	rgba.x = ((color & 0xff000000) >> 24) / 255; 
-	rgba.y = ((color & 0x00ff0000) >> 16) / 255;
-	rgba.z = ((color & 0x0000ff00) >> 8) / 255;
+	rgba.x = (float)((color & 0xff000000) >> 24) / 255; 
+	rgba.y = (float)((color & 0x00ff0000) >> 16) / 255;
+	rgba.z = (float)((color & 0x0000ff00) >> 8) / 255;
 	return (rgba);
 }
